Stage Code Refactoring

+++ Entities/Concrete/Car.cs
@@ -11,8 +11,8 @@ namespace Entities.Concrete
         public int CarId { get; set; } 
         public int BrandId { get; set; }//Brand = Marka
         public int ColorId { get; set; }
-        public int ModelYear { get; set; }
-        public int DailyPrice { get; set; }// DailyPrice = Günlük fiyat
+        public short ModelYear { get; set; }
+        public decimal DailyPrice { get; set; }// DailyPrice = Günlük fiyat
         public string Description { get; set; }//Descrition = Açıklama
 
     }
